+++ Formula/harnessx.rb
@@ -2,28 +2,28 @@
 class Harnessx < Formula
   desc "Deterministic agentic engineering harness with policy gates, sensors, MCPs, hooks, sandboxed runs, and portable backups"
   homepage "https://github.com/rodolfopeixoto/harnessx"
-  version "0.17.0"
+  version "0.18.0"
   license "MIT"
 
   on_macos do
     on_arm do
-      url "https://github.com/rodolfopeixoto/harnessx/releases/download/v0.17.0/harness-darwin-arm64.tar.gz"
-      sha256 "27c34cd573c404668723fd82f57991dcd312105de825fd74e5a14512b7eae107"
+      url "https://github.com/rodolfopeixoto/harnessx/releases/download/v0.18.0/harness-darwin-arm64.tar.gz"
+      sha256 "8eb00ad79bfc299f03138f783ea7c97d70c21ddc42b57c5b39a463036370fd4c"
     end
     on_intel do
-      url "https://github.com/rodolfopeixoto/harnessx/releases/download/v0.17.0/harness-darwin-amd64.tar.gz"
-      sha256 "f1c042a72e396367f86abf6772e92058ea169c149083d328da4974777c25f5c1"
+      url "https://github.com/rodolfopeixoto/harnessx/releases/download/v0.18.0/harness-darwin-amd64.tar.gz"
+      sha256 "72cec0d00fe6781039708fe0e84d3c165132cdd6f8af473a653b2857ac10f2f6"
     end
   end
 
   on_linux do
     on_arm do
-      url "https://github.com/rodolfopeixoto/harnessx/releases/download/v0.17.0/harness-linux-arm64.tar.gz"
-      sha256 "694a9354c1fb3d94181e280c1b8837a47adbb205c6c319631328c05cc9470250"
+      url "https://github.com/rodolfopeixoto/harnessx/releases/download/v0.18.0/harness-linux-arm64.tar.gz"
+      sha256 "554b7d76863b2efb7800020af384c91fc051853018ece264deea6d9df7213178"
     end
     on_intel do
-      url "https://github.com/rodolfopeixoto/harnessx/releases/download/v0.17.0/harness-linux-amd64.tar.gz"
-      sha256 "649207b6fb91437191da718a56ccafb64f3a9742d0494fd626476c7f62cf1ae2"
+      url "https://github.com/rodolfopeixoto/harnessx/releases/download/v0.18.0/harness-linux-amd64.tar.gz"
+      sha256 "95c03881f6c0415015d9ab19281357c437d4d97f9c457dcfd9809ac9907b57c0"
     end
   end